Throwing checked exception on missing attribute, refactoring of tests, simpler getOrDefault method.
