+++ src/test/java/com/github/piotrkot/json/JsonObjTest.java
@@ -30,10 +30,11 @@ import java.io.ByteArrayInputStream;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
 import javax.json.Json;
-import org.cactoos.iterable.IterableOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
 /**
  * Test for JSON object.
@@ -42,6 +43,12 @@ import org.junit.Test;
  * @checkstyle ClassDataAbstractionCoupling (2 lines)
  */
 public final class JsonObjTest {
+    /**
+     * Expected exception.
+     */
+    @Rule
+    public final ExpectedException exp = ExpectedException.none();
+
     /**
      * Should create JSON object.
      * @throws Exception When fails.
@@ -132,84 +139,32 @@ public final class JsonObjTest {
     }
 
     /**
+     * Should not fetch JSON object attribute.
+     * @throws Exception When fails.
+     */
+    @Test
+    public void shouldNotGetObjAttr() throws Exception {
+        this.exp.expect(
+            Matchers.allOf(
+                Matchers.isA(JsonException.class),
+                Matchers.hasProperty(
+                    "message",
+                    Matchers.is("attribute name \"missing\" not found")
+                )
+            )
+        );
+        new JsonObj().get("missing");
+    }
+
+    /**
      * Should fetch JSON object default attribute.
      * @throws Exception When fails.
      */
     @Test
     public void shouldGetObjDefAttr() throws Exception {
         MatcherAssert.assertThat(
-            new JsonObj().getOrDefault("miss", new Vstr("X")).value(),
+            new JsonObj().get("miss", new Vstr("X")).value(),
             Matchers.is("X")
-        );
-    }
-
-    /**
-     * Should transfer JSON object.
-     * Suffix "str" attribute with 'Z', and multiply "num" attribute by 2.
-     * <pre>{"str":"A","num":1} => {"str":"AZ","num":2}</pre>
-     * @throws Exception When fails.
-     */
-    @Test
-    public void shouldTransferObj() throws Exception {
-        MatcherAssert.assertThat(
-            new JsonObj(
-                new JsonObj(
-                    new JsonObj.Attr("str", new Vstr("A")),
-                    new JsonObj.Attr("num", new Vnum(1))
-                ),
-                new Change.Attr(
-                    "st.", str -> new Vstr(((String) str.value()).concat("Z"))
-                ),
-                new Change.Attr(
-                    ".*um",
-                    num -> new Vnum(((Number) num.value()).intValue() * 2)
-                )
-            ).jsonValue().toString(),
-            Matchers.is("{\"str\":\"AZ\",\"num\":2}")
-        );
-    }
-
-    /**
-     * Should transfer JSON object by filling missing entries.
-     * Fills "text" attribute with 'message' when missing.
-     * <pre>{"s":\"ss\"} => {"s":\"ss\","text":"text"}</pre>
-     * @throws Exception When fails.
-     */
-    @Test
-    public void shouldFillMissingObj() throws Exception {
-        MatcherAssert.assertThat(
-            new JsonObj(
-                new JsonObj(
-                    new JsonObj.Attr("s", new Vstr("ss"))
-                ),
-                new Change.AttrMiss("text", new Vstr("message")),
-                new Change.AttrDel("none")
-            ).jsonValue().toString(),
-            Matchers.is("{\"s\":\"ss\",\"text\":\"message\"}")
-        );
-    }
-
-    /**
-     * Should transfer JSON object by deleting given attribute.
-     * Deletes "del" attribute.
-     * <pre>{"del":\"priv\","keep":\"pub\"} => {"keep":\"pub\"}</pre>
-     * @throws Exception When fails.
-     */
-    @Test
-    @SuppressWarnings("PMD.AvoidDuplicateLiterals")
-    public void shouldDelAttrObj() throws Exception {
-        MatcherAssert.assertThat(
-            new JsonObj(
-                new JsonObj(
-                    new JsonObj.Attr("del", new Vstr("priv")),
-                    new JsonObj.Attr("keep", new Vstr("pub"))
-                ),
-                new IterableOf<>(
-                    new Change.AttrDel("del"),
-                    new Change.AttrMiss("keep", new Vstr("xxx"))
-                )
-            ).jsonValue().toString(),
-            Matchers.is("{\"keep\":\"pub\"}")
         );
     }
 }
